Refactor affect analyzers into modular package
