+++ src/diaremot/affect/analyzers/__init__.py
@@ -1,0 +1,54 @@
+"""Component analyzers for the affect processing stack."""
+
+from .common import (
+    EmotionOutputs,
+    GOEMOTIONS_LABELS,
+    SER8_LABELS,
+    IntentResult,
+    SpeechEmotionResult,
+    TextEmotionResult,
+    VadEmotionResult,
+    default_intent_result,
+    default_speech_result,
+    default_text_result,
+    default_vad_result,
+    json_dumps,
+    normalize_backend,
+    normalize_intent_label,
+    resolve_component_dir,
+    resolve_model_dir,
+    select_first_existing,
+)
+from .intent import IntentAnalyzer, resolve_intent_model_dir
+from .speech import OnnxAudioEmotion, SpeechEmotionAnalyzer
+from .text import HfTextEmotionFallback, OnnxTextEmotion, TextEmotionAnalyzer
+from .vad import OnnxVADEmotion, VadEmotionAnalyzer
+
+__all__ = [
+    "EmotionOutputs",
+    "GOEMOTIONS_LABELS",
+    "SER8_LABELS",
+    "IntentResult",
+    "SpeechEmotionResult",
+    "TextEmotionResult",
+    "VadEmotionResult",
+    "default_intent_result",
+    "default_speech_result",
+    "default_text_result",
+    "default_vad_result",
+    "json_dumps",
+    "normalize_backend",
+    "normalize_intent_label",
+    "resolve_component_dir",
+    "resolve_model_dir",
+    "select_first_existing",
+    "IntentAnalyzer",
+    "resolve_intent_model_dir",
+    "OnnxAudioEmotion",
+    "SpeechEmotionAnalyzer",
+    "HfTextEmotionFallback",
+    "OnnxTextEmotion",
+    "TextEmotionAnalyzer",
+    "OnnxVADEmotion",
+    "VadEmotionAnalyzer",
+]
